Refactor apps/classifications
